manually retrive Error due to out-of-tree module
Because this module is out-of-tree, Rust for some reason refuses to
acknowledge that `Error: From<VarInt>` is satisfied. This does not occur
when the module is in-tree. Simply including the trait definitions from
s2n-quic doesn't work either. This patch provides a fix which allows
s2n-quic-h3 to still compile out-of-tree; it does not need to be
forwarded back into s2n-quic, as the issue does not exist in-tree.

Signed-off-by: Amy Parker <amy@amyip.net>